refactor(apply): repackage and change url
