+++ src/main/kotlin/apply/ui/admin/evaluation/EvaluationsView.kt
@@ -1,7 +1,8 @@
-package apply.ui.admin
+package apply.ui.admin.evaluation
 
 import apply.application.EvaluationService
 import apply.domain.evaluation.dto.EvaluationResponse
+import apply.ui.admin.BaseLayout
 import com.vaadin.flow.component.Component
 import com.vaadin.flow.component.grid.Grid
 import com.vaadin.flow.component.html.H1
@@ -16,10 +17,8 @@ import support.createDeleteButtonWithDialog
 import support.createPrimaryButton
 import support.createPrimarySmallButton
 
-@Route(value = "admin/evaluation", layout = BaseLayout::class)
-class EvaluationView(
-    private val evaluationService: EvaluationService
-) : VerticalLayout() {
+@Route(value = "admin/evaluations", layout = BaseLayout::class)
+class EvaluationsView(private val evaluationService: EvaluationService) : VerticalLayout() {
     init {
         add(createTitle(), createButton(), createGrid())
     }
