+++ src/main/kotlin/apply/ui/admin/selections/SelectionsView.kt
@@ -1,7 +1,8 @@
-package apply.ui.admin
+package apply.ui.admin.selections
 
 import apply.application.RecruitmentService
 import apply.domain.recruitment.Recruitment
+import apply.ui.admin.BaseLayout
 import com.vaadin.flow.component.Component
 import com.vaadin.flow.component.UI
 import com.vaadin.flow.component.button.Button
